% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_names.R
\name{get_var_names}
\alias{get_var_names}
\title{Utility for identifying variables of different sorts}
\usage{
get_var_names(data, type = c("quantitative", "categorical", "discrete",
  "all"), max_levels = 10L, at_least_n = 0L)
}
\arguments{
\item{data}{a data table}

\item{type}{quantitative or categorical}

\item{max_levels}{integer specifying how many levels categorical variables}

\item{at_least_n}{minimum number of not-missing values in the variable
can have to be included in the list}
}
\description{
Returns the names of variables in \code{data} that meet certain criteria:
}
\details{
type "quantitative" -- any variable that's numeric
type "categorical" -- any variable thats a factor, a character, or a logical. Must have no nore than
\code{max_levels} levels.
type "discrete" -- either quantitative or categorical, having no more than \code{max_levels} distinct values
type "all" -- anything with enough non-missing values
}
\examples{
get_var_names(mtcars, type = "discrete", max_levels = 3)

}
