% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_as_plot.R
\name{gf_characters}
\alias{gf_characters}
\title{Plot out text strings as individual items}
\usage{
gf_characters(object = NULL, gformula = row ~ col, textdf)
}
\description{
To support interactive text, this function generates a plot where each character
has a known position
}
