% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_components.R
\name{main_display}
\alias{main_display}
\title{Layout utilities for little apps}
\usage{
main_display(sample_sizes = c(5, 10, 20, 50, 100, 500, 1000),
  brush = brushOpts("horiz_line", direction = "y", fill = NA),
  response_vars = NULL, explan_vars = "None", covars = "None",
  multiple_covariates = FALSE)
}
\arguments{
\item{sample_sizes}{numerical vector giving allowable sample sizes}

\item{brush}{the brush options for the plot in shiny (see \code{shiny::brushOpts()})}

\item{response_vars}{character vector giving names of response variables}

\item{explan_vars}{character vector giving names of response variables}

\item{covars}{character vector giving names of response variables}

\item{multiple_covariates}{Logical. If \code{TRUE}, allow multiple covariates to be selected.}
}
\value{
A shiny tag that can be displayed as one entity.
}
\description{
The shiny UI collection returned by this function has \code{input$} handles
that can be used in the server logic.
\itemize{
\item \code{plot} the target for \code{output$plot <- renderPlot()}
\item \code{response} the \code{input$} component holding the name of the response variable.
\item \code{explan} similarly for an explanatory variable
\item \code{covar} similarly for covariates. This might be a vector of names if \code{multiple_covariates = TRUE}
\item \code{seed} seed to generate a new sample
\item \code{explain} the target for \code{output$explain <- renderHTML()}
\item \code{codebook} the target for \code{output$codebook <- renderText()}
}
}
