% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_policy.R
\name{point_alpha}
\alias{point_alpha}
\title{Set transparency appropriately for a scatter plot}
\usage{
point_alpha(n)
}
\arguments{
\item{n}{the number of points to be displayed}
}
\description{
The transparency (alpha) value should be decreased slightly as
a greater number of points is displayed
}
