% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive.R
\name{react_calc}
\alias{react_calc}
\title{Create reactive-style functions that can be run outside of shiny}
\usage{
react_calc(stmts, shiny = FALSE, env = parent.frame())
}
\arguments{
\item{stmts}{bracketed code of the same sort used in \code{shiny::reactive()}}

\item{shiny}{flag indicated whether a shiny-reactive function or a mock-reactive
function should be produced.}
}
\description{
Rseplacement for \code{shiny::reactive()}  that makes it
easy to shift between the reactive style in shiny and a mock-reactive style
that can be run in an ordinary session. In the reactive style, values used
in reactive functions are often produced by other reactive functions.
}
\examples{
input <- list(samp_n = 15)
get_data <- react_calc({head(mtcars, input$samp_n, shiny = FALSE)})
make_model <- react_calc({lm(mpg ~ hp, data = get_data())})
find_rsquared <- react_calc({summary(make_model())$r.squared})

find_rsquared()  # the ultimate calculation


}
