% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribbon_for_violin.R
\name{ribbon_for_violin}
\alias{ribbon_for_violin}
\title{Create data for a horizontal ribbon that mimics a violin plot}
\usage{
ribbon_for_violin(var, groups = NULL, vpos = 0, vwidth = 0.4,
  bw = "nrd0", adjust = 1)
}
\arguments{
\item{var}{a vector of numbers: the variable to be plotted}

\item{groups}{a numerical vector specifying the vertical position of each group.}

\item{vpos}{a number specifying the y-position of the center of the violin. See details.}

\item{vwidth}{the maximum half-height of the violin of a box in which
the violin will be drawn. The actual width will be set so that the violin
occupies 1/4 of the box.}

\item{bw}{a bandwidth specified per \code{stats::density()}}

\item{adjust}{a number that can be used to adjust the bandwidth}
}
\description{
gf_violin() wants the quantitative variable to be on the y axis, but
sometimes you want a horizontal violin on top of other plots that can't be rotated,
for instance if they have a density plot or histogram.
}
\details{
If there are multiple groups, then vpos should be a numeric vector specifying
the position of each group's ribbon. If it's a single number, that number will be repeated
for all levels of the group.
}
\examples{
# all together
ribbon_data <- ribbon_for_violin(mtcars$mpg, vpos = 3)
gf_ribbon(top + bottom ~ x, data = ribbon_data)
# several groups
ribbon_data <- ribbon_for_violin(mtcars$mpg, as.numeric(mtcars$cyl))
gf_ribbon(top + bottom ~ x, data = ribbon_data, fill = ~ group)
}
