% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_app.R
\name{select_app}
\alias{select_app}
\title{Select one of the available apps and run it.}
\usage{
select_app()
}
\description{
For convenience only, this function will display a list of the available apps and let you run
one.
}
\details{
In actual use, you would display the app in a learnr tutorial or another Rmd document containing
narrative describing the app.
}
