% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_littleapp_server.R
\name{standard_littleapp_server}
\alias{standard_littleapp_server}
\title{Standard server function for a Little App}
\usage{
standard_littleapp_server(session, input, output, V)
}
\arguments{
\item{session}{the session argument from the usual app \code{server()} function}

\item{input}{the input argument from the usual app \code{server()} function.}

\item{output}{the output argument from the usual app \code{server()} function}

\item{V}{the reactive value holding the state}
}
\description{
This function provides the standard set of services for a Little App
}
\details{
The argument \code{V} is a reactive value created \emph{outside} of this function.
It holds the state which is the main means of communication among the controls
and the output displays.
}
\examples{
\dontrun{
# Your app's server function will look like this
server <- function(session, input, output) {
  V <- reactiveValues() # shared state
  standard_littleapp_server(session, input, output)
  # you can add other reactives, e.g.
  custom_1 <- reactive({
    # internal logic
  })
  output$custom2 <- renderImage({
    # internal logic
  })
}

}

}
