% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scale.R
\name{StatPREP_color_scale}
\alias{StatPREP_color_scale}
\title{Color scale for categorical values}
\usage{
StatPREP_color_scale(levels, annots = NULL, annot_colors = "red")
}
\arguments{
\item{levels}{vector of character strings with names of data levels}

\item{annots}{vector of character strings with names of "annotation" levels}

\item{annot_colors}{name or names of colors to use for the annotations. Default: \code{"red"}. These will be
used in the same order as \code{annots} and repeated as needed.}
}
\description{
Generates a color scale for categorical values.
}
