% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatPREPshiny.R
\name{StatPREPshiny}
\alias{StatPREPshiny}
\title{Shiny apps for inclusion in other documents}
\description{
In order to include Shiny apps in a learnr tutorial, the apps
have been written as Rmd documents. These can then be included in a tutorial
by using the \code{child=system.file("name_of_app.Rmd", package = "StatPREPshiny")} option in
a chunk.
}
\details{
In the app Rmd files, the chunks containing the server part of
the app should be marked as \code{context="server"}.
}
