% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulers.R
\name{v_ruler}
\alias{v_ruler}
\title{Add rulers to a plot}
\usage{
v_ruler(object = NULL, gformula = NULL, data = NULL, description = "sd",
  nticks = 3, color = "black", alpha = 1, size = 0.5, width = NULL)
}
\description{
This is still under development. In particular, I'm not sure how to pass
the arguments \code{color}, \code{alpha}, and so on to the gf_functions being called.
Currently, they are fixed. The \code{width} argument is particularly important.
}
\examples{
rule_data <- data.frame(center = c(3,5), length = c(1.5, 2.7), x = c("A", "B"))
v_ruler(center + length ~ x, data = rule_data, width = NULL)

}
