% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_policy.R
\name{verbatim_colors}
\alias{verbatim_colors}
\title{Color scale for named colors
Use this so that named colors get shown as the color itself}
\usage{
verbatim_colors()
}
\description{
Color scale for named colors
Use this so that named colors get shown as the color itself
}
\examples{
scale_y_continuous( name = "probabability", breaks = seq(0, 1, by = 0.2),
                    verbatim_colors())

}
